+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Issue #300: Add clean Radar map state
+
 ### Changed
 
 - Issue #275: Unify yaml configuration files